LPD-75534 Move method from PageExperienceUtil to SegmentsExperienceUtil
