+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/src/main/java/com/liferay/headless/admin/site/internal/resource/v1_0/util/LayoutUtil.java
@@ -1231,7 +1231,7 @@ public class LayoutUtil {
 
 				SegmentsExperienceServiceUtil.updateSegmentsExperiencePriority(
 					actualSegmentsExperience.getSegmentsExperienceId(),
-					PageExperienceUtil.getPriority(
+					SegmentsExperienceUtil.getPriority(
 						pageExperience.getKey(), layout,
 						pageExperience.getPriority()));
 			}
